requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
appdirs==1.4.4            # via black, virtualenv
argcomplete==1.12.2       # via nox
attrs==20.3.0             # via pytest
black==20.8b1             # via -r requirements-dev.in
click==7.1.2              # via black
colorlog==4.6.2           # via nox
distlib==0.3.1            # via virtualenv
filelock==3.0.12          # via virtualenv
flake8-print==4.0.0       # via -r requirements-dev.in
flake8==3.8.4             # via -r requirements-dev.in, flake8-print
iniconfig==1.1.1          # via pytest
mccabe==0.6.1             # via flake8
mypy-extensions==0.4.3    # via black
nox==2020.8.22            # via -r requirements-dev.in
packaging==20.7           # via pytest
pathspec==0.8.1           # via black
pluggy==0.13.1            # via pytest
py==1.9.0                 # via nox, pytest
pycodestyle==2.6.0        # via flake8, flake8-print
pyflakes==2.2.0           # via flake8
pyparsing==2.4.7          # via packaging
pytest==6.1.2             # via -r requirements-dev.in
regex==2020.11.13         # via black
six==1.15.0               # via flake8-print, virtualenv
toml==0.10.2              # via black, pytest
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
virtualenv==20.2.2        # via nox
